wip: implementing cell stream
